[FE] refactor: BoxModal, ProfilePage 컴포넌트로 이동
- ProfileTechStackBox에서 생성했던 BoxModal을 ProfilePage 컴포넌트에서 생성하여 전역으로 관리하도록 변경(추후 다른 ProfileBox에서 모달의 element만 바꿔서 재사용이 가능하도록 하기위함)
- issue (#65)
